puml export

+++ database/puml.cpp
@@ -43,16 +43,31 @@ namespace database
     }
 
     std::string Puml::generate_puml(std::vector<Node> &nodes,
-                                    std::vector<Link> &links)
+                                    std::vector<Link> &links,
+                                    std::string default_path)
     {
         std::string result= Poco::UUIDGenerator().createRandom().toString();
 
         _futures[result]=std::async(std::launch::async,
-                    [nodes,links,result](){
-                        
+                    [nodes,links,result,default_path](){
                         auto escape_string = [](std::string &str){
                             std::string res;
-                            std::transform(std::begin(str),std::end(str),std::back_inserter(res),[](char c){ return c==','?'_':c;});
+                            std::transform(std::begin(str),std::end(str),std::back_inserter(res),
+                            [](char c){
+                                switch(c){
+                                    case '\'': return ' ';
+                                    case '\"': return ' ';
+                                    case '/':  return '_';
+                                    case '\\': return '_';
+                                    case ':': return '_';
+                                    case ',': return '_';
+                                    case '&': return '_';
+                                    case '(': return '_';
+                                    case ')': return '_';
+                                    default:
+                                        return c;
+                                };
+                            });
                             return res;
                         };
                         // create directory
@@ -61,7 +76,13 @@ namespace database
 
                         // create puml
                         std::ofstream   file;
-                        std::string     file_name="puml/"+result+".puml";
+                        std::string     file_name;
+                        std::map<std::string,std::string> codes;
+
+                        if(default_path.empty()) file_name="puml/"+result+".puml";
+                                            else file_name=default_path+".puml";
+                        
+                        std::cout << "Generating:" << file_name << std::endl;
                         file.open(file_name, std::ios::binary);
 
                         file    << "@startuml" << std::endl;
@@ -77,16 +98,19 @@ namespace database
                                 if(n.get()["type"]=="Подразделение") type="Person";
                             }
                             std::string code = "code_"+std::to_string(++i);
-                            if(n.get().find("code")!=std::end(n.get())) code = n.get()["code"];
+                            codes[ n.get()["code"]] = code;
+                            /*if(n.get().find("code")!=std::end(n.get())) {
+                                code = n.get()["code"];
+                            }*/
 
-                            code = escape_string(code);
+                            //code = escape_string(code);
                             std::string name = code;
                             if(n.get().find("name")!=std::end(n.get())) name = n.get()["name"];
 
                             std::string description;
                             for(auto & [k1,k2]: n.get()) description += k1 +":"+k2+" ";
 
-                            file << "component \"" << name << "\" as " << escape_string(code) << std::endl;
+                            file << "component \"" << escape_string(name) << "\" as " << codes[ n.get()["code"]] << "" << std::endl;
                             // file    << type << "(" << code <<",\"" << name <<"\",\"" << description<< "\")" << std::endl;
                         }
 
@@ -113,7 +137,7 @@ namespace database
                                 target_node_code = l.target_node_code();
                                 name += l.get()["name"]+"; ";
                             }
-                            file << escape_string(source_node_code) << " --> " << escape_string(target_node_code) << " : " << name << std::endl;
+                            file << "" << codes[source_node_code] << " --> " << codes[target_node_code] << " : \"" << escape_string(name) << "\"" << std::endl;
                             //file    << "Rel(" << escape_string(source_node_code) <<"," << escape_string(target_node_code) <<",\"" << name <<"\")" << std::endl;
                         }
 
@@ -129,6 +153,8 @@ namespace database
                         // run puml generator
                         std::string command="java -jar plantuml.jar "+file_name;
                         ssystem(command.c_str());
+
+                        if(default_path.empty())
                         std::experimental::filesystem::remove(file_name);
                         
                     });
